Updated files to match final submission
