+++ app.py
@@ -5,8 +5,6 @@ from time import *
 from models import *
 from helper import *
 
-
-#print("DONE")
 curruser = ''
 vcl=[]  #VariableCardList
 vcll=-1  #VariableCardList-Last
@@ -19,10 +17,6 @@ def index():
 		return render_template('index.html')
 	elif request.method == "POST":
 		uname = request.form["uid"]
-		print(uname)
-		#uid=users.query.filter_by(username=uname).first()
-		#curruid=uid.uid
-		#print(curruid)
 		global curruser
 		curruser = uname
 		return redirect(url_for('dash', uname=curruser))
@@ -30,13 +24,11 @@ def index():
 
 @app.route('/user/<uname>', methods=["GET", "POST"])
 def dash(uname):
-	#print("dash",uname)
 	name = users.query.filter_by(username=uname).first()
 	dks = decks.query.all()
 
 	
 	for deck in dks:
-		#lrt(deck.did)
 		if deck.cid != '':
 			scores(deck.did)
 		
@@ -49,21 +41,22 @@ def deck_update(id):
 
 		dks = decks.query.filter_by(did=id).first()
 		cds = cards.query.all()
-		#print(id)
 		return render_template('editdeck.html', deck=dks, cards=cds)
 
 	if request.method == "POST":
 		global curruser
-		#d_id=request.form["did"]
 		dname = request.form["dname"]
 		ddesc = request.form["ddesc"]
 		cardlist = request.form.getlist("cardselect")
+
+		if len(cardlist)<=0:
+			return redirect('/deckemptyerror')  
+
+
 		x = ''
 		for i in cardlist:
 			x = x + str(i) + ','
 		x = x[:-1]
-		#print(x)
-		#print(dname,ddesc)
 		x = decks.query.filter_by(did=id).update(
 		    dict(deck_name=dname, deck_description=ddesc, cid=x))
 		db.session.commit()
@@ -83,23 +76,18 @@ def deck_delete(id):
 def deck_create():
 	global curruser
 	if request.method == "GET":
-		#global curruser
 		cds = cards.query.all()
 		return render_template('newdeck.html', cards=cds)
 
 	if request.method == "POST":
-		#global curruser
 		d_id = request.form["did"]
 		dname = request.form["dname"]
 		ddesc = request.form["ddesc"]
 		cardlist = request.form.getlist("cardselect")
-		#cardlist = request.args.get('cardselect')
-		print(cardlist, type(cardlist))
 		x = ''
 		for i in cardlist:
 			x = x + str(i) + ','
 		x = x[:-1]
-		print(x)
 		z = decks(did=d_id, deck_name=dname, deck_description=ddesc, cid=x)
 		db.session.add(z)
 		db.session.commit()
@@ -125,23 +113,22 @@ def newcard():
 		cobv = request.form["cobv"]
 		crev = request.form["crev"]
 		decklist = request.form.getlist("deckselect")
-		#cardlist = request.args.get('cardselect')
-		#print(cardlist,type(cardlist))
-		x = ''
-		for i in decklist:
-			x = x + str(i) + ','
-		x = x[:-1]
-		q = convint(x)
-		for i in q:
-			w = decks.query.filter_by(did=i).first()
-			dold = w.cid
-			if dold == '':
-				dold = dold + str(c_id)
-			else:
-				dold = dold + ',' + str(c_id)
-			c = decks.query.filter_by(did=i).update(dict(cid=dold))
-			db.session.commit()
-			scores(i)
+		if len(decklist)>0: 
+			x = ''
+			for i in decklist:
+				x = x + str(i) + ','
+			x = x[:-1]
+			q = convint(x)
+			for i in q:
+				w = decks.query.filter_by(did=i).first()
+				dold = w.cid
+				if dold == '':
+					dold = dold + str(c_id)
+				else:
+					dold = dold + ',' + str(c_id)
+				c = decks.query.filter_by(did=i).update(dict(cid=dold))
+				db.session.commit()
+				scores(i)
 		z = cards(cid=c_id, obverse=cobv, reverse=crev)
 		db.session.add(z)
 		db.session.commit()
@@ -158,15 +145,33 @@ def test():
 	if request.method == "GET":
 		did = request.args.get("did", type=int)
 		cl = request.args.get("cid", type=str)
-		print(did, cl)
+
+		if len(cl)<=0:
+			return redirect('/deckerror')  
+
 		clist = convint(cl)
-		print('did:', did, 'clist:', clist)
 		clist.reverse()
 		vcl = clist
 		vcll=vcl[-1]
 		cdi = did
 
 		return redirect('/deck/review/start')
+
+@app.route('/deckerror', methods=["GET", "POST"])
+def deckerror():
+	if request.method=="GET":
+		return render_template('deckerror.html')
+	if request.method=="POST":
+		return redirect(url_for('dash',uname=curruser))
+
+
+@app.route('/deckemptyerror', methods=["GET", "POST"])
+def deckemptyerror():
+	if request.method=="GET":
+		return render_template('deckemptyerror.html')
+	if request.method=="POST":
+		return redirect(url_for('dash',uname=curruser))
+
 @app.route('/deck/review/start', methods=["GET", "POST"])
 def reviewproc():
 	global curruser
@@ -179,17 +184,11 @@ def reviewproc():
       
 			cl = request.args.get("diff", type=int)
 			
-			print('cl:',type(cl))
-			print('vcl:',vcl)
-			print('vcll:',vcll)
 			x = vcl.pop()
 			
 			w = decks.query.filter_by(did=cdi).first()
 			q = cards.query.filter_by(cid=x).first()
 
-			print('x:',x,'w.deckname:', w.deck_name,'obverse:',q.obverse,'reverse:',q.reverse)
-			print('vcl',vcl)
-			
 			if cl is not None:
 				c=cards.query.filter_by(cid=vcll).update(dict(score=cl))
 				db.session.commit()
@@ -230,7 +229,6 @@ def reviewproc():
 
 @app.route('/exitdeck',methods=["GET"])
 def exitdeck():
-	#random mumbo jumbo about updating last reviewed time
 	return redirect(url_for('dash',uname=curruser))
 
 
@@ -243,8 +241,6 @@ def deletecard():
 
 	if request.method == "POST":
 		cardlist = request.form.getlist("cardselect")
-		#cardlist = request.args.get('cardselect')
-		print(cardlist, type(cardlist))
 		x = ''
 		for i in cardlist:
 			x = x + str(i) + ','
@@ -258,9 +254,6 @@ def deletecard():
 	
 	return redirect(url_for('dash', uname=curruser))
 
-
-
-
-
+	
 if __name__ == '__main__':
-	app.run(debug=True, host='0.0.0.0', port='80')
+	app.run(debug=False, host='0.0.0.0', port='80')
